backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pymysql
email-validator
dnspython
cryptography
